+++ c2ctask/src/c2c.ts
@@ -1,0 +1,7 @@
+
+/*
+* Interface for the CodeToCloud Decorator task.
+*/
+export interface ICodeToCloud {
+    run(): any;
+}
